requirements.txt
mysql-connector-python==8.0.17
requests==2.21.0

